Change Camera::uri meaning
Two fields changed their meaning:

1) Optional `onvif_addr` is removed.
To access camera we used `onvif_addr` for IP cams
and `uri` for USB/CSI cams, so now it becomes more
consistent - we always use `uri`. For IP cameras `uri`
now is ONVIF device management service URL. For
USB/CSI cameras it is left as it was before: local file 
name.

2) Since we change `uri` field meaning we are loosing 
the field that was used for storing IP camera's RTSP URI.
So I've created one for this purpose: `rtsp_uri`.